refactor(spx-gui): move loadimage function out of painter component
