+++ spx-gui/src/components/editor/common/painter/utils/SvgImporter.vue
